Moved hashlists from utils to accumulators package
