+++ packages/protocol/src/index.ts
@@ -12,12 +12,12 @@ export * from "./model/transaction/ValueOption";
 export * from "./model/MethodPublicOutput";
 export * from "./model/RuntimeLike";
 export * from "./model/AppliedStateTransitionBatch";
-export * from "./utils/ProvableHashList";
-export * from "./utils/PrefixedProvableHashList";
-export * from "./utils/MinaPrefixedProvableHashList";
-export * from "./utils/ProvableReductionHashList";
+export * from "./accumulators/ProvableHashList";
+export * from "./accumulators/PrefixedProvableHashList";
+export * from "./accumulators/MinaPrefixedProvableHashList";
+export * from "./accumulators/ProvableReductionHashList";
 export * from "./utils/utils";
-export * from "./utils/FieldOptions";
+export * from "./model/FieldOptions";
 export * from "./prover/accumulators/StateTransitionReductionList";
 export * from "./prover/accumulators/AppliedBatchHashList";
 export * from "./prover/accumulators/WitnessedRootHashList";
