+++ packages/protocol/src/prover/accumulators/TransactionHashList.ts
@@ -1,6 +1,6 @@
 import { Field } from "o1js";
 
-import { DefaultProvableHashList } from "../../utils/ProvableHashList";
+import { DefaultProvableHashList } from "../../accumulators/ProvableHashList";
 
 export class TransactionHashList extends DefaultProvableHashList<Field> {
   public constructor(commitment: Field = Field(0)) {
